src/requirements.txt
# Core Python Version
python-dotenv>=1.0.0

# Web Framework
flask>=2.0.0
flask-cors>=4.0.0

# Web scraping
selenium>=4.10.0
beautifulsoup4>=4.12.2
scrapy>=2.11.0
requests>=2.31.0

# Data handling and preprocessing
pandas>=2.0.3
numpy>=1.25.0
scikit-learn>=1.3.0
spacy>=3.7.0

# Dimensionality reduction and embedding visualization
umap-learn>=0.5.4

# Embedding models and vector DB
sentence-transformers>=2.2.2
pinecone-client>=2.2.4

# Vector Databases
pinecone-client>=3.0.0
chromadb>=0.5.3

# Web frameworks (you can use either Flask or FastAPI)
flask>=3.0.0
fastapi>=0.110.0
pydantic>=2.4.0
uvicorn>=0.23.0

# For automation, scheduling, and background tasks
apscheduler>=3.10.1

# AWS SDK for Python
boto3>=1.28.0

# For MCP (Model Context Protocol) integration if needed later
openai>=1.13.3
tiktoken>=0.7.0

# Optional but helpful tools
python-multipart>=0.0.6
gunicorn>=21.2.0
